requirements.txt
Django>=1.7.0,<1.8
django_polymorphic
django-multiupload
markdown
django-celery
django-suit
Pillow
django_compressor
pygments
pyPdf
requests
xmltodict
djangorestframework
drf-extensions
furl
django-mptt

# Dev
django-extensions
ipython
honcho

# Test
mock
